move dto and entity to model package and remove management block from application.yml
