--- src/main/java/com/instahipsta/harCRUD/dto/Transferable.java
@@ -1,4 +1,0 @@
-package com.instahipsta.harCRUD.dto;
-
-public interface Transferable {
-}
