--- src/main/java/com/instahipsta/harCRUD/entity/Entityable.java
@@ -1,4 +1,0 @@
-package com.instahipsta.harCRUD.entity;
-
-public interface Entityable {
-}
