+++ src/main/java/com/instahipsta/harCRUD/mapper/AbstractMapper.java
@@ -1,7 +1,7 @@
 package com.instahipsta.harCRUD.mapper;
 
-import com.instahipsta.harCRUD.dto.Transferable;
-import com.instahipsta.harCRUD.entity.Entityable;
+import com.instahipsta.harCRUD.model.dto.Transferable;
+import com.instahipsta.harCRUD.model.entity.Entityable;
 import org.modelmapper.Converter;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
