+++ src/main/java/com/instahipsta/harCRUD/mapper/HarMapper.java
@@ -1,7 +1,7 @@
 package com.instahipsta.harCRUD.mapper;
 
-import com.instahipsta.harCRUD.dto.HarDTO;
-import com.instahipsta.harCRUD.entity.Har;
+import com.instahipsta.harCRUD.model.dto.HarDTO;
+import com.instahipsta.harCRUD.model.entity.Har;
 import org.springframework.stereotype.Component;
 
 @Component
