+++ src/main/java/com/instahipsta/harCRUD/mapper/Mapper.java
@@ -1,7 +1,7 @@
 package com.instahipsta.harCRUD.mapper;
 
-import com.instahipsta.harCRUD.dto.Transferable;
-import com.instahipsta.harCRUD.entity.Entityable;
+import com.instahipsta.harCRUD.model.dto.Transferable;
+import com.instahipsta.harCRUD.model.entity.Entityable;
 
 public interface Mapper<E extends Entityable, D extends Transferable> {
