+++ src/main/java/com/instahipsta/harCRUD/model/dto/Transferable.java
@@ -1,0 +1,4 @@
+package com.instahipsta.harCRUD.model.dto;
+
+public interface Transferable {
+}
