+++ src/main/java/com/instahipsta/harCRUD/model/entity/Entityable.java
@@ -1,0 +1,4 @@
+package com.instahipsta.harCRUD.model.entity;
+
+public interface Entityable {
+}
