+++ src/main/java/com/instahipsta/harCRUD/model/entity/TestProfile.java
@@ -1,4 +1,4 @@
-package com.instahipsta.harCRUD.entity;
+package com.instahipsta.harCRUD.model.entity;
 
 import javax.persistence.*;
 import java.util.List;
