+++ src/main/java/com/instahipsta/harCRUD/repository/RequestRepo.java
@@ -1,6 +1,6 @@
 package com.instahipsta.harCRUD.repository;
 
-import com.instahipsta.harCRUD.entity.Request;
+import com.instahipsta.harCRUD.model.entity.Request;
 import org.springframework.data.repository.CrudRepository;
 
 public interface RequestRepo extends CrudRepository<Request, Long> {
