+++ src/main/java/com/instahipsta/harCRUD/repository/TestProfileRepo.java
@@ -1,6 +1,6 @@
 package com.instahipsta.harCRUD.repository;
 
-import com.instahipsta.harCRUD.entity.TestProfile;
+import com.instahipsta.harCRUD.model.entity.TestProfile;
 import org.springframework.data.repository.CrudRepository;
 
 public interface TestProfileRepo extends CrudRepository<TestProfile, Long> {
