+++ src/main/java/com/instahipsta/harCRUD/service/TestProfileService.java
@@ -1,8 +1,8 @@
 package com.instahipsta.harCRUD.service;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.instahipsta.harCRUD.entity.Request;
-import com.instahipsta.harCRUD.entity.TestProfile;
+import com.instahipsta.harCRUD.model.entity.Request;
+import com.instahipsta.harCRUD.model.entity.TestProfile;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
