+++ src/main/java/com/instahipsta/harCRUD/service/impl/HarServiceImpl.java
@@ -2,8 +2,8 @@ package com.instahipsta.harCRUD.impl;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.instahipsta.harCRUD.dto.HarDTO;
-import com.instahipsta.harCRUD.entity.Har;
+import com.instahipsta.harCRUD.model.dto.HarDTO;
+import com.instahipsta.harCRUD.model.entity.Har;
 import com.instahipsta.harCRUD.mapper.HarMapper;
 import com.instahipsta.harCRUD.repository.HarRepo;
 import com.instahipsta.harCRUD.service.HarService;
