+++ src/main/resources/application.yml
@@ -65,15 +65,6 @@ spring:
     enabled: true
     change-log: classpath*:db.changelog/db.changelog-master-dev.yaml
 
-management:
-  endpoints:
-    enabled-by-default: false
-  endpoint:
-    health:
-      enabled: true
-    info:
-      enabled: true
-
 file:
   downloads: ./downloads
