+++ src/test/java/com/instahipsta/harCRUD/dto/HarDtoTest.java
@@ -1,9 +1,9 @@
 package com.instahipsta.harCRUD.dto;
 
+import com.instahipsta.harCRUD.model.dto.HarDTO;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
