+++ src/test/java/com/instahipsta/harCRUD/entity/TestProfileTest.java
@@ -1,5 +1,7 @@
 package com.instahipsta.harCRUD.entity;
 
+import com.instahipsta.harCRUD.model.entity.Request;
+import com.instahipsta.harCRUD.model.entity.TestProfile;
 import com.instahipsta.harCRUD.service.RequestService;
 import com.instahipsta.harCRUD.service.TestProfileService;
 import org.junit.Assert;
