+++ src/test/java/com/instahipsta/harCRUD/impl/HarServiceTest.java
@@ -1,7 +1,7 @@
 package com.instahipsta.harCRUD.impl;
 
-import com.instahipsta.harCRUD.entity.Har;
-import com.instahipsta.harCRUD.dto.HarDTO;
+import com.instahipsta.harCRUD.model.entity.Har;
+import com.instahipsta.harCRUD.model.dto.HarDTO;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
